langchain_openai_demo/pyproject.toml
[tool.poetry]
name = "langchain-openai-demo"
version = "0.1.0"
description = ""
authors = ["kirin <kirin.chen1001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "1.1"
langchain = "^0.1.17"
python-dotenv = "^1.0.1"
google-search-results = "^2.4.2"
numexpr = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
